use pylint to improve style of code
The code is currently not very Pythonic (e.g. in use of camel case
function and variable names, how lists are iterated).

This commit adds the use of the tox tool to run pylint across the
code and updates it to reach a reasonable quality level. Reaching
10/10 is not really feasible wihout much more extensive refactoring,
so some errors have been disabled.