add a title to each Bingo game
To make it easier to tell which ticket and track listing file belongs
to each Bingo game, add a title each Bingo ticket, the track listing
file and the game results file.

The title is automatically generated from the directory name from
the Clips folder. If songs are selected from multiple folders, no
title is used.