Add audio playback to the GUI
To make the clip generation process a bit easier, play the selected
part of an audio file when the title is double clicked.

Installing the pyaudio package becomes an optional package, which
allows a clip to be played. Use of pyaudio is recommended because its
API allows playback to be aborted.