use multiple threads to search directories
The searching for songs can be quite slow if many MP3 files have
to be parsed. This commit modifies the directory class to create
tasks to check each file and tasks to search subdirectories.
These tasks are executed in multiple threads using a
concurrent.futures.ThreadPoolExecutor.

There is a minor regression in how the progress percentage is
calculated as the total number of songs and directories is unknown
when the searching starts. As each directory is discovered and the
total task count becomes know, the progress percentage estimation
improves.