use ffmpeg for MP3 editing
Given that pydub uses ffmpeg for its MP3 encoding, a better approach
is to use ffmpeg directly. This has the advantage of making generation
faster and uses less memory, as there is no need to load the entire
audio samples into Python memory.

It also allows audio files to be played using "ffplay" without the need
to install he pyaudio library, which can be tricky on Windows.