WIP: add HTTP server to allow users to select and view tickets
With the world in lock down, we need a way to play Music Bingo
online.

This commit adds an HTTP server that serves a simple HTML /
JavaScript application. This app allows a user to create an account,
view available Bingo games and then select one or more tickets from
each game.

The ticket can be played online by clicking on each song or the
ticket can be downloaded as a PDF file.