remove schema version table
Using a table to store Schema versions has not worked well as
it is brittle and can lead to situations where the table is
out of sync with the migrations that have been applied.

It is possible to detect the schema of each table directly from
the table, so this is hopefully a more robust way to handle
schema migrations.