when importing game tracks, add directories if not found in database
When importing a gameTracks file, it will create Song entries if
they are not found in the database. However this can fail if the
directory is not found in the database.

This commit causes directories to be created for songs if they are
not found in the database. If a directory is used by more than one
song and the artists of those songs do not match, the artist field
of this new directory is changed to "Various Artists".