make CodePipeline optional and fix CodeDeploymentGroup bucket
Make the creation of CodeDeploy and CodePipeline resources optional.

If these are enabled, the CodeDeploymentGroup create step was failing
because it tries to deploy a musicbingo-prod.tar.gz file from an S3
bucket that has only just been created.

Once the stack has been created, the musicbingo-prod.tar.gz file
needs to be uploaded to the S3 bucket, which will then cause
the deploy process to trigger.