move artist to its own table
To improve database normalisation, replace the artist string field
with an Artist table. This reduces database size, makes searches
easier, removes duplication and makes it easier to modify.

Unpicking this Schema change does require quite a lot of changes,
mostly to unit tests and anything connected to the Song class.