try harder to find existing songs when importing games
When importing a game, it should be using songs that already
exist in the database where possible. However this was not
the case and in many cases it would create duplicate entries
for a song.

This commit updates the import logic to check if a song already
exists in the database, and only create a new song if not found.
When adding a song, a new directory bingo+games/{game_id} is
created as the parent for these new songs.