assign ref_id values after all songs have been scanned
If a new directory is added to the Clips directory, the
dd_available_songs_to_treeview() function will fail to
add the new songs because it causes duplicate ref_id
values.

To fix this bug, wait until the clips directory has
been searched and then assign IDs to each directory
and to each song.