dont link to track listing if using a guest account.
Guest users are not able to see the track listing, but the past games
page was providing a link to the track listing page. This commit
removes the link and adds a message to suggest to the guest that
they need to register a full account to see the track listing.