implement include_artist option
The include_artist game option can be used to disable the display
of a track's artist. This option was not supported in the client.

This commit causes the ticket display to only show the arist if
enabled and also adds an option to the ModifyGame component to
modify the include_artist option.