allow re-generation of a Bingo game
The musicbingo.generator.generate() function was not correctly
handling the case of being called when a game already exists.
This caused the function to throw an exception because it was
generating sqlalchemy errors for a duplicate game ID.

This commit changes that to check if the game already exists
and only adding a new game to the database if it does not exist.