dont use ref_id to search for songs when loading a game from database
Using the ref_id to find songs from the list of available songs
doesn't work correctly as it can choose completely different songs
from the ones when the game was originally generated.

This commit changes the logic to first search for the directory
and when found, search for the song by title in that directory.