show when each row of each ticket completes
For games where there is a prize for the first person to complete a
horizontal line, it is quite useful for the results page to show
for each ticket when each row becomes complete.

This commit updates the generator to calculate when each row
becomes complete and adds this information into the results PDF
page.