move ticket backgrounds from tickets.scss to the auto-generated CSS file
The tickets.scss file was defining the theme-based backgrounds for each
ticket icon on the ticket picking page. This is an issue because it
can become out of sync with the themes defined in the Python app.

This commit moves the icon images into the public directory and uses
the auto-generated CSS file to define CSS rules for each ticket icon.