allow privacy policy settings to be loaded when user not logged in
A user of the website should be able to read the privacy policy
without needing to have logged into the site. This commit modifies
the server API to return the privacy policy settings for a request
from a user who has not logged in, or when they are not an admin
user.

Only an admin user is allowed to access the full settings.