use asFragment to generate object for snapshots
Using react-test-renderer to create the JavaScript object was causing
warnings to be generated for each snapshot test, because there are two
versions of the act() method, which are not compatible. This commit
switches to just using '@testing-library/react' which has an
asFrament() function that provides a JS object that can be used for
snapshot testing.