add unit test for SettingsForm component
This commit modifies Input fields to wrap the title of each field
in a label element. This made it easier to find the field using the
DOM testing library.

The component test has a fault where the values of boolean fields
are sometimes showing as having the wrong value. When using the
application, the checkbox fields have the correct value, which
implies that the issue is with the test rather than the SettingsField
component.