refactor AdminGameActions to become a Component with child components
Prior to this commit, the AdminGameActions class was extended by
each component that wanted to make use of it. This is not a typical
React pattern. It is better to have a <AdminGameActions> component
that can be placed around other components, which are rendered as
child elements of AdminGameActions.render()