dont manually set options when loading a game from the database
When a game is loaded from the database, the options are updated
to match those used to generate that game. The app was going this
itself by directly setting properties in the options object. This
doesn't work because it doesn't handle enumerations. Using the
update() function in the Options class handles this for the app.