update API proxy to work with Node 18
When upgrading to Node v18, the connection from the development
client to the Python server stopped working. The webpack proxy
was unable to connect to the server, evev though it was correctly
responding to requests.

The commit updates to a slightly different method of configuring the
proxy that works with the current LTS version of Nodejs.