refactor test helpers
The test helpers had become rather unwieldy, especially the mock
API server. This commit creates a new class MockBingoServer that
provides a fake version of the Python server that implements some
of the logic of the server.

Additionally, all of the test code has been moved into the tests
directory, so that is separated from the source code in the src
directory.