change mock server to require bearer tokens in requests
To make the tests more closely match how the real API server
works, change the mock server to require that a bearer token
is included in HTTP requests. The user database in the mock
server is checked for a matching access token. A matching token
is required for any protected API request.