fix infinite loop when refreshing access token fails
If the access token refreshing, the refreshToken() function is called,
which in turn uses the same makeApiRequest() function to call the
JSON API to refresh the access token. If that refresh fails, it was
causing an infinite loop of trying to use the refresh token to update
the refresh token.